Separate modules for timestamps, positions, and data_ext
Remove parser and util modules. Move
functionality to specific modules.

Transfer encode/decode functionality to
Timestamp and Position objects to simplify
the actual packet objects (which can
work transparently with various formats)
