+++ tests/test_decode_regression.py
@@ -13,21 +13,19 @@ import pytest
 
 from kiss3.ax25 import Frame
 
-from aprs3 import parser
 from aprs3.classes import (
-    CourseSpeed,
     DataType,
-    DFS,
     InformationField,
     ItemReport,
     Message,
     ObjectReport,
-    PHG,
     PositionReport,
-    RNG,
     StatusReport,
 )
 from aprs3.constants import PositionFormat, TimestampFormat
+from aprs3.data_ext import CourseSpeed, DFS, PHG, RNG
+from aprs3.position import Position
+from aprs3.timestamp import Timestamp
 
 
 @pytest.fixture(autouse=True)
@@ -37,7 +35,7 @@ def fixed_now(monkeypatch):
     def new_now(*args, **kwargs):
         return datetime.datetime(2022, 5, 23, 23, 59, tzinfo=datetime.timezone.utc)
 
-    with mock.patch("aprs3.parser.datetime") as mock_datetime:
+    with mock.patch("aprs3.timestamp.datetime") as mock_datetime:
         mock_datetime.now.side_effect = new_now
         mock_datetime.strptime.side_effect = (
             lambda *args, **kw: datetime.datetime.strptime(*args, **kw)
@@ -54,14 +52,15 @@ def fixed_now(monkeypatch):
                 raw=b"=4704.13N/12242.73W[241/055/A=-00053 Mike ",
                 data_type=DataType.POSITION_W_O_TIMESTAMP_MSG,
                 data=b"4704.13N/12242.73W[",
-                data_ext=CourseSpeed(course=241, speed=55),
                 comment=b"/A=-00053 Mike ",
-                timestamp=None,
-                lat=Decimal("47.06883333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-122.7121666666666666666666667"),
-                symbol_code=b"[",
-                altitude_ft=-53,
+                position=Position(
+                    lat=Decimal("47.06883333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-122.7121666666666666666666667"),
+                    symbol_code=b"[",
+                    data_ext=CourseSpeed(course=241, speed=55),
+                    altitude_ft=-53,
+                ),
             ),
             id="position, uncompressed, without timestamp, course/speed, altitude, comment",
         ),
@@ -71,13 +70,14 @@ def fixed_now(monkeypatch):
                 raw=b"!4605.21N/12327.31W#PHG2830W2, ORn-N, Fill-in / NA7Q 14.3V 44.2F",
                 data_type=DataType.POSITION_W_O_TIMESTAMP,
                 data=b"4605.21N/12327.31W#",
-                data_ext=PHG(power_w=4, height_ft=2560, gain_db=3, directivity=0),
                 comment=b"W2, ORn-N, Fill-in / NA7Q 14.3V 44.2F",
-                timestamp=None,
-                lat=Decimal("46.08683333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-123.4551666666666666666666667"),
-                symbol_code=b"#",
+                position=Position(
+                    lat=Decimal("46.08683333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-123.4551666666666666666666667"),
+                    symbol_code=b"#",
+                    data_ext=PHG(power_w=4, height_ft=2560, gain_db=3, directivity=0),
+                ),
             ),
             id="position, uncompressed, without timestamp, phg, comment",
         ),
@@ -87,13 +87,13 @@ def fixed_now(monkeypatch):
                 raw=b"!4558.13NF12259.58W MEISSNER LOOKOUT N7QXO-7",
                 data_type=DataType.POSITION_W_O_TIMESTAMP,
                 data=b"4558.13NF12259.58W ",
-                data_ext=b"",
                 comment=b"MEISSNER LOOKOUT N7QXO-7",
-                timestamp=None,
-                lat=Decimal("45.96883333333333333333333333"),
-                sym_table_id=b"F",
-                long=Decimal("-122.993"),
-                symbol_code=b" ",
+                position=Position(
+                    lat=Decimal("45.96883333333333333333333333"),
+                    sym_table_id=b"F",
+                    long=Decimal("-122.993"),
+                    symbol_code=b" ",
+                ),
             ),
             id="position, uncompressed, without timestamp, offset data type, comment",
         ),
@@ -103,16 +103,19 @@ def fixed_now(monkeypatch):
                 raw=b"@202350z4658.39N/12308.29W-WX3in1Plus2.0 U=13.9V",
                 data_type=DataType.POSITION_W_TIMESTAMP_MSG,
                 data=b"4658.39N/12308.29W-",
-                data_ext=b"",
                 comment=b"WX3in1Plus2.0 U=13.9V",
-                timestamp=datetime.datetime(
-                    2022, 5, 20, 23, 50, tzinfo=datetime.timezone.utc
+                timestamp=Timestamp(
+                    timestamp_format=TimestampFormat.DayHoursMinutesZulu,
+                    timestamp=datetime.datetime(
+                        2022, 5, 20, 23, 50, tzinfo=datetime.timezone.utc
+                    ),
                 ),
-                timestamp_format=TimestampFormat.DayHoursMinutesZulu,
-                lat=Decimal("46.97316666666666666666666667"),
-                sym_table_id=b"/",
-                long=Decimal("-123.1381666666666666666666667"),
-                symbol_code=b"-",
+                position=Position(
+                    lat=Decimal("46.97316666666666666666666667"),
+                    sym_table_id=b"/",
+                    long=Decimal("-123.1381666666666666666666667"),
+                    symbol_code=b"-",
+                ),
             ),
             id="position, uncompressed, with timestamp, comment",
         ),
@@ -122,13 +125,14 @@ def fixed_now(monkeypatch):
                 raw=b"!4605.21N/12327.31W#RNG0125Foo comment",
                 data_type=DataType.POSITION_W_O_TIMESTAMP,
                 data=b"4605.21N/12327.31W#",
-                data_ext=RNG(range=125),
                 comment=b"Foo comment",
-                timestamp=None,
-                lat=Decimal("46.08683333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-123.4551666666666666666666667"),
-                symbol_code=b"#",
+                position=Position(
+                    lat=Decimal("46.08683333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-123.4551666666666666666666667"),
+                    symbol_code=b"#",
+                    data_ext=RNG(range=125),
+                ),
             ),
             id="position, uncompressed, without timestamp, rng, comment",
         ),
@@ -138,13 +142,16 @@ def fixed_now(monkeypatch):
                 raw=b"!4605.21N/12327.31W#DFS8745Foo comment",
                 data_type=DataType.POSITION_W_O_TIMESTAMP,
                 data=b"4605.21N/12327.31W#",
-                data_ext=DFS(strength_s=8, height_ft=1280, gain_db=4, directivity=225),
                 comment=b"Foo comment",
-                timestamp=None,
-                lat=Decimal("46.08683333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-123.4551666666666666666666667"),
-                symbol_code=b"#",
+                position=Position(
+                    lat=Decimal("46.08683333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-123.4551666666666666666666667"),
+                    symbol_code=b"#",
+                    data_ext=DFS(
+                        strength_s=8, height_ft=1280, gain_db=4, directivity=225
+                    ),
+                ),
             ),
             id="position, uncompressed, without timestamp, dfs, comment",
         ),
@@ -155,11 +162,13 @@ def fixed_now(monkeypatch):
                 data_type=DataType.POSITION_W_O_TIMESTAMP_MSG,
                 data=b"/7.oh/FIK-  #",
                 comment=b" Masen in Longview",
-                position_format=PositionFormat.Compressed,
-                sym_table_id=b"/",
-                lat=Decimal("46.17683224563301"),
-                long=Decimal("-122.98066816127016"),
-                symbol_code=b"-",
+                position=Position(
+                    position_format=PositionFormat.Compressed,
+                    sym_table_id=b"/",
+                    lat=Decimal("46.17683224563301"),
+                    long=Decimal("-122.98066816127016"),
+                    symbol_code=b"-",
+                ),
             ),
             id="position, compressed, without timestamp",
         ),
@@ -169,13 +178,15 @@ def fixed_now(monkeypatch):
                 raw=b"=/7.oh/FIK-7P# Masen in Longview",
                 data_type=DataType.POSITION_W_O_TIMESTAMP_MSG,
                 data=b"/7.oh/FIK-7P#",
-                data_ext=CourseSpeed(course=88, speed=36.23201216883807),
                 comment=b" Masen in Longview",
-                position_format=PositionFormat.Compressed,
-                sym_table_id=b"/",
-                lat=Decimal("46.17683224563301"),
-                long=Decimal("-122.98066816127016"),
-                symbol_code=b"-",
+                position=Position(
+                    position_format=PositionFormat.Compressed,
+                    sym_table_id=b"/",
+                    lat=Decimal("46.17683224563301"),
+                    long=Decimal("-122.98066816127016"),
+                    symbol_code=b"-",
+                    data_ext=CourseSpeed(course=88, speed=36.23201216883807),
+                ),
             ),
             id="position, compressed, with cs",
         ),
@@ -185,7 +196,6 @@ def fixed_now(monkeypatch):
                 raw=b":KF0JGS-7 :@3037755154 I love you 2!{M1383",
                 data_type=DataType.MESSAGE,
                 data=b"KF0JGS-7 :@3037755154 I love you 2!{M1383",
-                data_ext=b"",
                 comment=b"",
                 addressee=b"KF0JGS-7",
                 text=b"@3037755154 I love you 2!",
@@ -199,7 +209,6 @@ def fixed_now(monkeypatch):
                 raw=b":SMSGTE   :ackM1383",
                 data_type=DataType.MESSAGE,
                 data=b"SMSGTE   :ackM1383",
-                data_ext=b"",
                 comment=b"",
                 addressee=b"SMSGTE",
                 text=b"ackM1383",
@@ -213,7 +222,6 @@ def fixed_now(monkeypatch):
                 raw=b":BLN1     :Net Mondays 19:00 146.840- T100.0",
                 data_type=DataType.MESSAGE,
                 data=b"BLN1     :Net Mondays 19:00 146.840- T100.0",
-                data_ext=b"",
                 comment=b"",
                 addressee=b"BLN1",
                 text=b"Net Mondays 19:00 146.840- T100.0",
@@ -227,7 +235,6 @@ def fixed_now(monkeypatch):
                 raw=b">Oregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
                 data_type=DataType.STATUS,
                 data=b"Oregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
-                data_ext=b"",
                 comment=b"",
                 status=b"Oregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
             ),
@@ -239,10 +246,12 @@ def fixed_now(monkeypatch):
                 raw=b">232114zOregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
                 data_type=DataType.STATUS,
                 data=b"Oregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
-                data_ext=b"",
                 comment=b"",
-                timestamp=datetime.datetime(
-                    2022, 5, 23, 21, 14, tzinfo=datetime.timezone.utc
+                timestamp=Timestamp(
+                    timestamp_format=TimestampFormat.DayHoursMinutesZulu,
+                    timestamp=datetime.datetime(
+                        2022, 5, 23, 21, 14, tzinfo=datetime.timezone.utc
+                    ),
                 ),
                 status=b"Oregon Coast Repeater Group: WX: Rose Lodge, OR: www.ocrg.org:W7GC-5",
             ),
@@ -254,23 +263,26 @@ def fixed_now(monkeypatch):
                 raw=b";W7ZA     *232209z4657.26N/12348.18WrPmin1,Pmax11,147.160+ T88.5 W7ZA.ORG",
                 data_type=DataType.OBJECT,
                 data=b"4657.26N/12348.18Wr",
-                data_ext=b"",
                 comment=b"Pmin1,Pmax11,147.160+ T88.5 W7ZA.ORG",
                 name=b"W7ZA",
                 killed=False,
-                timestamp=datetime.datetime(
-                    2022,
-                    5,
-                    23,
-                    22,
-                    9,
-                    tzinfo=datetime.timezone.utc,
+                timestamp=Timestamp(
+                    timestamp_format=TimestampFormat.DayHoursMinutesZulu,
+                    timestamp=datetime.datetime(
+                        2022,
+                        5,
+                        23,
+                        22,
+                        9,
+                        tzinfo=datetime.timezone.utc,
+                    ),
                 ),
-                timestamp_format=TimestampFormat.DayHoursMinutesZulu,
-                lat=Decimal("46.95433333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-123.803"),
-                symbol_code=b"r",
+                position=Position(
+                    lat=Decimal("46.95433333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-123.803"),
+                    symbol_code=b"r",
+                ),
             ),
             id="object1",
         ),
@@ -280,14 +292,15 @@ def fixed_now(monkeypatch):
                 raw=b")AID #2!4903.50N/07201.75WAfirst aid",
                 data_type=DataType.ITEM,
                 data=b"4903.50N/07201.75WA",
-                data_ext=b"",
                 comment=b"first aid",
                 name=b"AID #2",
                 killed=False,
-                lat=Decimal("49.05833333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-72.02916666666666666666666667"),
-                symbol_code=b"A",
+                position=Position(
+                    lat=Decimal("49.05833333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-72.02916666666666666666666667"),
+                    symbol_code=b"A",
+                ),
             ),
             id="item, live, comment",
         ),
@@ -297,14 +310,16 @@ def fixed_now(monkeypatch):
                 raw=b")AID #2_4903.50N/07201.75WA042/000first aid",
                 data_type=DataType.ITEM,
                 data=b"4903.50N/07201.75WA",
-                data_ext=CourseSpeed(course=42, speed=0),
                 comment=b"first aid",
                 name=b"AID #2",
                 killed=True,
-                lat=Decimal("49.05833333333333333333333333"),
-                sym_table_id=b"/",
-                long=Decimal("-72.02916666666666666666666667"),
-                symbol_code=b"A",
+                position=Position(
+                    lat=Decimal("49.05833333333333333333333333"),
+                    sym_table_id=b"/",
+                    long=Decimal("-72.02916666666666666666666667"),
+                    symbol_code=b"A",
+                    data_ext=CourseSpeed(course=42, speed=0),
+                ),
             ),
             id="item, killed, course/speed, comment",
         ),
